requirements.txt
pytorch
torchvision
imgaug
scikit-learn
jupyter
matplotlib
numpy
pandas
Pillow
tqdm